Phase 3.3b/4.1: Extract gem + streets from engine; consolidate colors
Gem extraction — was buried inside 670-line engine.ts.
- New scene/components/gem/gem.ts owns createRootGem + the cached
  procedural glow texture. Engine just imports + composes.

Street extraction — engine.ts had street geometry, sidewalk + asphalt
material setup, and street-label painting, all mixed with the scene
orchestrator.
- New scene/components/streets/streets.ts: createStreetMesh,
  flatGroundMaterial, and the stadium/pill geometry builder.
- New scene/components/streets/streetLabels.ts: createStreetLabels +
  _buildLabelTexture.
- New scene/components/streets/streetPath.ts: createPathMesh
  (building-to-street connector strips, drawn at the same Y as
  sidewalks).

Engine deleted — it had become a thin composition layer that only
glued the component factories together. Inlined as _buildCityScene
inside scene/cityScene.ts so the lifecycle owner and the one-shot
scene builder live side by side. cityScene is the only caller.

Color consolidation — colors lived across scene/colors.ts,
scene/hsl.ts, scene/shaders/hsl.glsl, and scene/shaders/registerShaderChunks.ts.
- All four moved to scene/utils/color/.
- Old scene/shaders/ directory removed (now empty; sky/building/label
  shaders live with their components).
- scene/components/trees/colorInterp.ts stays put — it's
  tree-specific OkLab interpolation, not a shared utility.

Verified: tsc clean, all 1872 web tests pass.

PR #24 review items addressed (partial): thalida E (consolidate
color files), K (gem extraction).
