+++ web/scene/utils/color/registerShaderChunks.ts
